% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opdelte funktioner.R
\name{ALK_sequence}
\alias{ALK_sequence}
\title{Identification of ALK breakpoint bases}
\usage{
ALK_sequence(reads, basepairs = 20)
}
\arguments{
\item{reads}{Data.frame returned by EML4_ALK_detection().}

\item{basepairs}{Number of basepairs identified from the EML4-ALK fusion. Default = 20.}
}
\value{
If EML4-ALK is detected, returns a table of identified ALK basepairs with the number of corresponding reads for each sequence. Otherwise "No EML4-ALK was detected" is returned.
}
\description{
This function identifies the basepairs following the ALK breakpoint.
}
\examples{
sample1 <- ALK_sequence(EML4_ALK_detection(file = "sample1.bam", genome = "hg38", mates = 2), basepairs = 20)
sample2 <- ALK_sequence(EML4_ALK_detection(file = "sample2.bam", genome = "hg38", mates = 2), basepairs = 20)
}
