% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAfusion_functions.R
\name{break_position_depth}
\alias{break_position_depth}
\title{Read depth at breakpoint}
\usage{
break_position_depth(file, reads)
}
\arguments{
\item{file}{The name of the file which the data are to be read from.}

\item{reads}{\code{GAlignments} object returned by EML4_ALK_detection().}
}
\value{
If EML4-ALK is detected a single integer corresponding
to the read depth at the breakpoint is returned.
Otherwise "No EML4-ALK was detected" is returned
}
\description{
This function identifies the read depth at the basepair
before the breakpoint in EML4.
}
\examples{
H3122_bam <- system.file("extdata",
"H3122_EML4.bam",
package="DNAfusion")
HCC827_bam <-  system.file("extdata",
"HCC827_EML4.bam",
package="DNAfusion")

break_position_depth(file=H3122_bam,
                        EML4_ALK_detection(file=H3122_bam,
                                            genome="hg38",
                                            mates=2))
break_position_depth(file=HCC827_bam,
                        EML4_ALK_detection(file=HCC827_bam,
                                            genome="hg38",
                                            mates=2))
}
