% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opdelte funktioner.R
\name{EML4_ALK_analysis}
\alias{EML4_ALK_analysis}
\title{Complete EML4-ALK analysis}
\usage{
EML4_ALK_analysis(file, genome = "hg38", mates = 2, basepairs = 20)
}
\arguments{
\item{file}{Is the name of the BAM file.}

\item{genome}{Is the reference genome. Can be either hg38 or hg19. Default = "hg38".}

\item{mates}{Is the minimum number EML4-ALK mate pairs needed to be detected in order to call a variant. Default = 2.}

\item{basepairs}{Number of basepairs identified from the EML4-ALK fusion. Default = 20.}
}
\value{
A list object with clipped_reads corresponding to EML_ALK_detection(), last_EML4 corresponding to EML4_sequence(), first_ALK corresponding to ALK_sequence(), breakpoint corresponding to break_position(), and read_depth corresponding to break_position_depth().
}
\description{
This functions collects the results from the other functions of the package.
}
\examples{
sample1 <- EML4_ALK_analysis(file = "sample1.bam", genome = "hg38", mates = 2, basepairs = 20)
sample2 <- EML4_ALK_analysis(file = "sample2.bam", genome = "hg38", mates = 2, basepairs = 20)
}
