% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAfusion_functions.R
\name{break_position}
\alias{break_position}
\title{EML4-ALK breakpoint}
\usage{
break_position(reads)
}
\arguments{
\item{reads}{Data.frame returned by EML4_ALK_detection().}
}
\value{
If EML4-ALK is detected, returns a \code{table} of genomic positions
with the number of corresponding reads for each sequence.
Otherwise "No EML4-ALK was detected" is returned.
}
\description{
This function identifies the genomic position in EML4
where the breakpoint has happened.
}
\examples{
H3122_bam <- system.file("extdata",
"H3122_EML4.bam",
package="DNAfusion")
HCC827_bam <-  system.file("extdata",
"HCC827_EML4.bam",
package="DNAfusion")

break_position(EML4_ALK_detection(file=H3122_bam,
                                    genome="hg38",
                                    mates=2))
break_position(EML4_ALK_detection(file=HCC827_bam,
                                    genome="hg38",
                                    mates=2))
}
