% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opdelte funktioner.R
\name{break_position}
\alias{break_position}
\title{EML4-ALK breakpoint}
\usage{
break_position(reads)
}
\arguments{
\item{reads}{Data.frame returned by EML4_ALK_detection().}
}
\value{
If EML4-ALK is detected, returns a table of genomic positions with the number of corresponding reads for each sequence. Otherwise "No EML4-ALK was detected" is returned.
}
\description{
This function identifies the genomic position in EML4 where the breakpoint has happened.
}
\examples{
sample1 <- break_position(EML4_ALK_detection(file = "sample1.bam", genome = "hg38", mates = 2))
sample2 <- break_position(EML4_ALK_detection(file = "sample2.bam", genome = "hg38", mates = 2))
}
