% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALKfusiondiscover_functions.R
\name{break_position_depth}
\alias{break_position_depth}
\title{Read depth at breakpoint}
\usage{
break_position_depth(file, reads)
}
\arguments{
\item{file}{Is the name of the BAM file.}

\item{reads}{Data.frame returned by EML4_ALK_detection().}
}
\value{
If EML4-ALK is detected a single integer corresponding to the read depth at the breakpoint is returned. Otherwise "No EML4-ALK was detected" is returned
}
\description{
This function identifies the read depth at the basepair before the breakpoint in EML4.
}
\examples{
sample1 <- break_position_depth(file = "sample1.bam",EML4_ALK_detection(file = "sample1.bam", genome = "hg38", mates = 2))
sample2 <- break_position_depth(file = "sample2.bam",EML4_ALK_detection(file = "sample2.bam", genome = "hg38", mates = 2))
}
