% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opdelte funktioner.R
\name{EML4_ALK_detection}
\alias{EML4_ALK_detection}
\title{Detection of EML4-ALK variants}
\usage{
EML4_ALK_detection(file, genome = "hg38", mates = 2)
}
\arguments{
\item{file}{Is the name of the BAM file.}

\item{genome}{Is the reference genome. Can be either hg38 or hg19. Default = "hg38".}

\item{mates}{Is the minimum number EML4-ALK mate pairs needed to be detected in order to call a variant. Default = 2.}
}
\value{
If EML4-ALK is detected a data.frame with soft-clipped reads representing EML4-ALK is returned. Otherwise "No EML4-ALK was detected" is returned
}
\description{
This function looks for EML4-ALK mate pair reads in the BAM file.
}
\examples{
sample1 <- EML4_ALK_detection(file = "sample1.bam", genome = "hg38", mates = 2)
sample2 <- EML4_ALK_detection(file = "sample2.bam", genome = "hg38", mates = 2)
}
